loop-38.c: Correct testcase.
	* gcc.dg/tree-ssa/loop-38.c: Correct testcase.

	* tree-ssa-loop-niter.c (discover_iteration_bound_by_body_walk): Add
	missing sign to unsigned.

From-SVN: r193117

+++ gcc/tree-ssa-loop-niter.c
@@ -3161,7 +3161,7 @@ discover_iteration_bound_by_body_walk (struct loop *loop)
     }
 
   gcc_assert (latch_index >= 0);
-  if (latch_index < VEC_length (double_int, bounds))
+  if ((unsigned)latch_index < VEC_length (double_int, bounds))
     {
       if (dump_file && (dump_flags & TDF_DETAILS))
 	{
